+++ front-end/app/components/App.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Header from "./Header";
+
+function App() {
+    return (
+      <Header />
+    );
+  }
+  
+  export default App;
