+++ front-end/app/components/Header.js
@@ -4,7 +4,6 @@ import Typography from "@mui/material/Typography";
 import Grid from "@mui/material/Grid";
 import TextField from "@mui/material/TextField";
 import Box from "@mui/material/Box";
-import Profile from "./Profile"
 
 
 function Header() {
@@ -29,7 +28,7 @@ function Header() {
                     <TextField variant="filled" color="info" focused></TextField>
                 </Grid>
                 <Grid item xs="3">
-                    <Profile />
+                    <button>Profile</button>
                 </Grid>
             </Grid>
             </Toolbar>
