+++ front-end/package.json
@@ -13,8 +13,7 @@
   "dependencies": {
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
-    "@mui/icons-material": "^5.11.0",
-    "@mui/material": "^5.11.7",
+    "@mui/material": "^5.11.6",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
