--- src/front-end/app/components/App.js
@@ -1,8 +1,0 @@
-import React from "react";
-import Header from "./Header";
-
-function App() {
-  return <Header />;
-}
-
-export default App;
